+++ package.json
@@ -34,7 +34,8 @@
     "ethers": "^5.4.4",
     "hardhat": "=2.6.1",
     "lint-staged": "^10.0.10",
-    "prettier": "^2.0.2"
+    "prettier": "^2.0.2",
+    "react-identicons": "^1.2.5"
   },
   "scripts": {
     "dev": "next dev",
