+++ src/styles/main.css
@@ -1,0 +1,7 @@
+.canvas {
+    margin-left: .5rem;
+    margin-bottom: 5px;
+    display: inline;
+    border-radius: 1.125rem;
+    background-color: black;
+}
